+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.1.3] - 2020-02-10
 
+- Add `check-page-security` function
 - Add page security for `/gallery` route
 - Fix SSR book card page on mobile